Seeder créé, migrat créé

+++ database/migrations/2023_10_04_102309_create_formateurs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nom', 25);
             $table->string('prenom', 25);
-            $table->tinyInteger('age');
+            $table->integer('age');
             $table->timestamps();  
         });
     }
